requirements.txt
langchain==0.1.20
langchain_community==0.0.38
langchain_openai==0.1.7
openai==1.30.1
fastapi==0.111.0
pydantic==2.7.1
uvicorn==0.29.0
python-dotenv==1.0.1
